Change Project To A 'Museum'
While working on the RichText editor version for YIlfipaeg, I realised
it would't fullfill my original idea of  showing the Yilfipaeg's
components in a more dynamic way. So I choose to go another
direction.

Now, I inted to make a App to 'simulate' a museum. It will allow the
user to select the component he wants to try out, along with it's
modifiers.